+++ packages/paste-core/core-bundle/src/index.tsx
@@ -44,6 +44,7 @@ export * from '@twilio-paste/tabs';
 export * from '@twilio-paste/tabs-primitive';
 export * from '@twilio-paste/text';
 export * from '@twilio-paste/textarea';
+export * from '@twilio-paste/time-picker';
 export * from '@twilio-paste/toast';
 export * from '@twilio-paste/tooltip';
 export * from '@twilio-paste/tooltip-primitive';