+++ packages/paste-website/src/components/site-wrapper/sidebar/SidebarNavigation.tsx
@@ -244,7 +244,7 @@ const SidebarNavigation: React.FC<SidebarNavigationProps> = () => {
                 </SidebarAnchor>
               </SidebarNestedItem>
               <SidebarNestedItem>
-                <SidebarAnchor nested to={`${SidebarCategoryRoutes.PATTERNS}/notifications`}>
+                <SidebarAnchor nested to={`${SidebarCategoryRoutes.PATTERNS}/notifications-and-feedback`}>
                   Notifications and feedback
                 </SidebarAnchor>
               </SidebarNestedItem>
@@ -315,20 +315,18 @@ const SidebarNavigation: React.FC<SidebarNavigationProps> = () => {
                   Overview
                 </SidebarAnchor>
               </SidebarNestedItem>
-              {data.allPastePrimitive.edges
-                .filter(({node}) => node.status !== PackageStatus.BACKLOG)
-                .map(({node}) => {
-                  return (
-                    <SidebarNestedItem key={node.name}>
-                      <SidebarAnchor
-                        nested
-                        to={`${SidebarCategoryRoutes.PRIMITIVES}/${getNameFromPackageName(node.name)}`}
-                      >
-                        {getHumanizedNameFromPackageName(node.name)}
-                      </SidebarAnchor>
-                    </SidebarNestedItem>
-                  );
-                })}
+              {data.allPastePrimitive.edges.filter(filteredComponents).map(({node}) => {
+                return (
+                  <SidebarNestedItem key={node.name}>
+                    <SidebarAnchor
+                      nested
+                      to={`${SidebarCategoryRoutes.PRIMITIVES}/${getNameFromPackageName(node.name)}`}
+                    >
+                      {getHumanizedNameFromPackageName(node.name)}
+                    </SidebarAnchor>
+                  </SidebarNestedItem>
+                );
+              })}
             </SidebarNestedList>
           </DisclosurePrimitiveContent>
         </SidebarItem>